requirements.txt
praw
python-dotenv
google-generativeai
